Updated gobenchpress application with MR details
There is currently an issue with the go-chart library this application
depends upon - in the case of no range existing in the data provided to
a bar chart, an error is returned (which is created by fmt.Errorf) -
this makes the error difficult to compare against and handle this case
more gracefully.  I have submitted a merge request for the functionality
change, using `errors.New` and comparing against a `var` declared at the
top of the file - however this change will have to await that MR being
merged and released.