Added support for outputting all benchmarks
Previously, Go Benchpress would output the last benchmark provided within
the Go benchmark output - now, all benchmarks will be output.  This is
based on a file mask as the output filename, with '{}' being used to
indicate where the benchmark's name should be put in the resulting file.
To achieve the previous behaviour, simply omit '{}' within the output
switch.